Refactor Jekyll config, layouts and includes
Front Matter has been modiefied on almost all pages to reflect removal of editable elements. However, special pages are almost untouched.

Because of the naming clash of categories collection with the special meaning of categories in Jekyll was the collection renamed to filterable tags and categories in photos and user answers were changed to tags because that is what they are from Jekyll’s point of view.